requirements.txt
selene == 2.0.0b17
pytest == 7.2.2
allure-pytest == 2.13.0